+++ src/js/input.js
@@ -1,47 +1,46 @@
 'use strict';
-
 /**
- * Input helper functions.
+ * Form control (input, select, textarea) helper functions.
  */
 
 module.exports = {
     // Multiple nodes are limited to ones of the same input type (better implemented as JQuery plugin actually)
-    getWrapNodes: function( $inputNodes ) {
-        var type = this.getInputType( $inputNodes.eq( 0 ) );
-        return ( type === 'fieldset' ) ? $inputNodes : $inputNodes.closest( '.question, .calculation' );
+    getWrapNodes: function( $inputs ) {
+        var type = this.getInputType( $inputs.eq( 0 ) );
+        return ( type === 'fieldset' ) ? $inputs : $inputs.closest( '.question, .calculation' );
     },
     /** very inefficient, should actually not be used **/
-    getProps: function( $node ) {
-        if ( $node.length !== 1 ) {
+    getProps: function( $input ) {
+        if ( $input.length !== 1 ) {
             return console.error( 'getProps(): no input node provided or multiple' );
         }
         return {
-            path: this.getName( $node ),
-            ind: this.getIndex( $node ),
-            inputType: this.getInputType( $node ),
-            xmlType: this.getXmlType( $node ),
-            constraint: this.getConstraint( $node ),
-            calculation: this.getCalculation( $node ),
-            relevant: this.getRelevant( $node ),
-            readonly: this.getReadonly( $node ),
-            val: this.getVal( $node ),
-            required: this.getRequired( $node ),
-            enabled: this.isEnabled( $node ),
-            multiple: this.isMultiple( $node )
+            path: this.getName( $input ),
+            ind: this.getIndex( $input ),
+            inputType: this.getInputType( $input ),
+            xmlType: this.getXmlType( $input ),
+            constraint: this.getConstraint( $input ),
+            calculation: this.getCalculation( $input ),
+            relevant: this.getRelevant( $input ),
+            readonly: this.getReadonly( $input ),
+            val: this.getVal( $input ),
+            required: this.getRequired( $input ),
+            enabled: this.isEnabled( $input ),
+            multiple: this.isMultiple( $input )
         };
     },
-    getInputType: function( $node ) {
+    getInputType: function( $input ) {
         var nodeName;
-        if ( $node.length !== 1 ) {
+        if ( $input.length !== 1 ) {
             return ''; //console.error('getInputType(): no input node provided or multiple');
         }
-        nodeName = $node.prop( 'nodeName' ).toLowerCase();
+        nodeName = $input.prop( 'nodeName' ).toLowerCase();
         if ( nodeName === 'input' ) {
-            if ( $node.data( 'drawing' ) ) {
+            if ( $input.data( 'drawing' ) ) {
                 return 'drawing';
             }
-            if ( $node.attr( 'type' ).length > 0 ) {
-                return $node.attr( 'type' ).toLowerCase();
+            if ( $input.attr( 'type' ).length > 0 ) {
+                return $input.attr( 'type' ).toLowerCase();
             }
             return console.error( '<input> node has no type' );
 
@@ -55,36 +54,36 @@ module.exports = {
             return console.error( 'unexpected input node type provided' );
         }
     },
-    getConstraint: function( $node ) {
-        return $node.attr( 'data-constraint' );
+    getConstraint: function( $input ) {
+        return $input.attr( 'data-constraint' );
     },
-    getRequired: function( $node ) {
+    getRequired: function( $input ) {
         // only return value if input is not a table heading input
-        if ( $node.parentsUntil( '.or', '.or-appearance-label' ).length === 0 ) {
-            return $node.attr( 'data-required' );
+        if ( $input.parentsUntil( '.or', '.or-appearance-label' ).length === 0 ) {
+            return $input.attr( 'data-required' );
         }
     },
-    getRelevant: function( $node ) {
-        return $node.attr( 'data-relevant' );
+    getRelevant: function( $input ) {
+        return $input.attr( 'data-relevant' );
     },
-    getReadonly: function( $node ) {
-        return $node.is( '[readonly]' );
+    getReadonly: function( $input ) {
+        return $input.is( '[readonly]' );
     },
-    getCalculation: function( $node ) {
-        return $node.attr( 'data-calculate' );
+    getCalculation: function( $input ) {
+        return $input.attr( 'data-calculate' );
     },
-    getXmlType: function( $node ) {
-        if ( $node.length !== 1 ) {
+    getXmlType: function( $input ) {
+        if ( $input.length !== 1 ) {
             return console.error( 'getXMLType(): no input node provided or multiple' );
         }
-        return $node.attr( 'data-type-xml' );
+        return $input.attr( 'data-type-xml' );
     },
-    getName: function( $node ) {
+    getName: function( $input ) {
         var name;
-        if ( $node.length !== 1 ) {
+        if ( $input.length !== 1 ) {
             return console.error( 'getName(): no input node provided or multiple' );
         }
-        name = $node.attr( 'data-name' ) || $node.attr( 'name' );
+        name = $input.attr( 'data-name' ) || $input.attr( 'name' );
         return name || console.error( 'input node has no name' );
     },
     /**
@@ -93,30 +92,30 @@ module.exports = {
      * NOTE: this function should be used sparingly, as it is CPU intensive!
      * TODO: simplify this function by looking for nodes with same CLASS on wrapNode
      *
-     * @param  {jQuery} $node The jQuery-wrapped input element
+     * @param  {jQuery} $input The jQuery-wrapped input element
      * @return {number}       The index
      */
-    getIndex: function( $node ) {
+    getIndex: function( $input ) {
         var inputType;
         var name;
         var $wrapNode;
         var $wrapNodesSameName;
 
-        if ( $node.length !== 1 ) {
+        if ( $input.length !== 1 ) {
             return console.error( 'getIndex(): no input node provided or multiple' );
         }
 
-        inputType = this.getInputType( $node );
-        name = this.getName( $node );
-        $wrapNode = this.getWrapNodes( $node );
+        inputType = this.getInputType( $input );
+        name = this.getName( $input );
+        $wrapNode = this.getWrapNodes( $input );
 
-        if ( inputType === 'radio' && name !== $node.attr( 'name' ) ) {
+        if ( inputType === 'radio' && name !== $input.attr( 'name' ) ) {
             $wrapNodesSameName = this.getWrapNodes( this.form.view.$.find( '[data-name="' + name + '"]' ) );
         }
         // fieldset.or-group wraps fieldset.or-repeat and can have same name attribute!)
-        else if ( inputType === 'fieldset' && $node.hasClass( 'or-repeat' ) ) {
+        else if ( inputType === 'fieldset' && $input.hasClass( 'or-repeat' ) ) {
             $wrapNodesSameName = this.getWrapNodes( this.form.view.$.find( '.or-repeat[name="' + name + '"]' ) );
-        } else if ( inputType === 'fieldset' && $node.hasClass( 'or-group' ) ) {
+        } else if ( inputType === 'fieldset' && $input.hasClass( 'or-group' ) ) {
             $wrapNodesSameName = this.getWrapNodes( this.form.view.$.find( '.or-group[name="' + name + '"]' ) );
         } else {
             $wrapNodesSameName = this.getWrapNodes( this.form.view.$.find( '[name="' + name + '"]' ) );
@@ -124,54 +123,56 @@ module.exports = {
 
         return $wrapNodesSameName.index( $wrapNode );
     },
-    isMultiple: function( $node ) {
-        return ( this.getInputType( $node ) === 'checkbox' || $node.attr( 'multiple' ) !== undefined ) ? true : false;
+    isMultiple: function( $input ) {
+        return ( this.getInputType( $input ) === 'checkbox' || $input.attr( 'multiple' ) !== undefined ) ? true : false;
     },
-    isEnabled: function( $node ) {
-        return !( $node.prop( 'disabled' ) || $node.parentsUntil( '.or', '.disabled' ).length > 0 );
+    isEnabled: function( $input ) {
+        return !( $input.prop( 'disabled' ) || $input.parentsUntil( '.or', '.disabled' ).length > 0 );
     },
-    getVal: function( $node ) {
+    getVal: function( $input ) {
         var inputType;
         var values = [];
         var name;
 
-        if ( $node.length !== 1 ) {
+        if ( $input.length !== 1 ) {
             return console.error( 'getVal(): no inputNode provided or multiple' );
         }
-        inputType = this.getInputType( $node );
-        name = this.getName( $node );
+        inputType = this.getInputType( $input );
+        name = this.getName( $input );
 
         if ( inputType === 'radio' ) {
-            return this.getWrapNodes( $node ).find( 'input:checked' ).val() || '';
+            return this.getWrapNodes( $input ).find( 'input:checked' ).val() || '';
         }
         // checkbox values bug in jQuery as (node.val() should work)
         if ( inputType === 'checkbox' ) {
-            this.getWrapNodes( $node ).find( 'input[name="' + name + '"]:checked' ).each( function() {
+            this.getWrapNodes( $input ).find( 'input[name="' + name + '"]:checked' ).each( function() {
                 values.push( this.value );
             } );
             return values;
         }
-        return $node.val() || '';
+        return $input.val() || '';
     },
-    setVal: function( name, index, value ) {
-        var $inputNodes;
-        var type;
-        var curVal;
-        var d;
-        var ds;
-
-        index = index || 0;
-
+    find: function( name, index ) {
+        var attr = 'name';
         if ( this.getInputType( this.form.view.$.find( '[data-name="' + name + '"]:not(.ignore)' ).eq( 0 ) ) === 'radio' ) {
-            type = 'radio';
-            $inputNodes = this.getWrapNodes( this.form.view.$.find( '[data-name="' + name + '"]' ) ).eq( index ).find( '[data-name="' + name + '"]:not(.ignore)' );
+            attr = 'data-name';
+        }
+        return this.getWrapNodes( this.form.view.$.find( '[' + attr + '="' + name + '"]' ) ).eq( index ).find( '[' + attr + '="' + name + '"]:not(.ignore)' ).eq( 0 );
+    },
+    setVal: function( $input, value ) {
+        var $inputs;
+        var type = this.getInputType( $input );
+        var $question = this.getWrapNodes( $input );
+        var name = this.getName( $input );
+
+        if ( type === 'radio' ) {
+            $inputs = $question.find( '[data-name="' + name + '"]:not(.ignore)' );
         } else {
             // why not use this.getIndex?
-            $inputNodes = this.getWrapNodes( this.form.view.$.find( '[name="' + name + '"]' ) ).eq( index ).find( '[name="' + name + '"]:not(.ignore)' );
-            type = this.getInputType( $inputNodes.eq( 0 ) );
+            $inputs = $question.find( '[name="' + name + '"]:not(.ignore)' );
 
             if ( type === 'file' ) {
-                $inputNodes.eq( 0 ).attr( 'data-loaded-file-name', value );
+                $input.attr( 'data-loaded-file-name', value );
                 // console.error('Cannot set value of file input field (value: '+value+'). If trying to load '+
                 //  'this record for editing this file input field will remain unchanged.');
                 return false;
@@ -179,8 +180,8 @@ module.exports = {
 
             if ( type === 'date' || type === 'datetime' ) {
                 // convert current value (loaded from instance) to a value that a native datepicker understands
-                // TODO test for IE, FF, Safari when those browsers start including native datepickers
-                value = this.form.model.node( name, index ).convert( value, type );
+                // TODO: test for IE, FF, Safari when those browsers start including native datepickers
+                value = this.form.model.types[ type ].convert( value );
             }
 
             if ( type === 'time' ) {
@@ -190,12 +191,12 @@ module.exports = {
                     // Strip the thousands of a second, because most browsers fail to parse such a time.
                     // Add a space before the timezone offset to satisfy some browsers.
                     // For IE11, we also need to strip the Left-to-Right marks \u200E...
-                    ds = new Date().toLocaleDateString( 'en', {
+                    var ds = new Date().toLocaleDateString( 'en', {
                         month: 'short',
                         day: 'numeric',
                         year: 'numeric'
                     } ).replace( /\u200E/g, '' ) + ' ' + value.replace( /(\d\d:\d\d:\d\d)(\.\d{1,3})(\s?((\+|-)\d\d))(:)?(\d\d)?/, '$1 GMT$3$7' );
-                    d = new Date( ds );
+                    var d = new Date( ds );
                     if ( d.toString() !== 'Invalid Date' ) {
                         value = d.getHours().toString().pad( 2 ) + ':' + d.getMinutes().toString().pad( 2 );
                     } else {
@@ -205,7 +206,7 @@ module.exports = {
             }
         }
 
-        if ( this.isMultiple( $inputNodes.eq( 0 ) ) === true ) {
+        if ( this.isMultiple( $input ) === true ) {
             value = value.split( ' ' );
         } else if ( type === 'radio' ) {
             value = [ value ];
@@ -213,16 +214,16 @@ module.exports = {
 
         // Trigger an 'inputupdate' event which can be used in widgets to update the widget when the value of its 
         // original input element has changed **programmatically**.
-        if ( $inputNodes.length ) {
-            curVal = this.getVal( $inputNodes.eq( 0 ) );
+        if ( $inputs.length ) {
+            var curVal = this.getVal( $input );
             if ( curVal === undefined || curVal.toString() !== value.toString() ) {
-                $inputNodes.val( value );
+                $inputs.val( value );
                 // don't trigger on all radiobuttons/checkboxes
-                $inputNodes.eq( 0 ).trigger( 'inputupdate.enketo' );
+                $inputs.eq( 0 ).trigger( 'inputupdate.enketo' );
             }
         }
 
-        return $inputNodes[ 0 ];
+        return $inputs[ 0 ];
     },
     validate: function( $input ) {
         return this.form.validateInput( $input );